Move *.tsx files to components/
- Moved all .tsx files to components folder:
    - Footer.tsx
    - Header.tsx
    - Layout.tsx
- Moved logo.png to assets folder.
