+++ frontend/src/components/Layout.tsx
@@ -1,6 +1,6 @@
 import {Outlet} from "react-router"
-import Header from "../Header/Header"
-import Footer from "../Footer/Footer"
+import Header from "./Header"
+import Footer from "./Footer"
 import {AppShell} from "@mantine/core"
 import {useDisclosure} from "@mantine/hooks"
 import styles from "./Layout.module.css"
